+++ modules/mssql/main.tf
@@ -25,8 +25,9 @@ locals {
   databases = { for db in var.additional_databases : db.name => db }
   users     = { for u in var.additional_users : u.name => u }
 
-  retained_backups = lookup(var.backup_configuration, "retained_backups", null)
-  retention_unit   = lookup(var.backup_configuration, "retention_unit", null)
+  retained_backups      = lookup(var.backup_configuration, "retained_backups", null)
+  retention_unit        = lookup(var.backup_configuration, "retention_unit", null)
+  connector_enforcement = var.connector_enforcement ? "REQUIRED" : "NOT_REQUIRED"
 }
 
 resource "random_id" "suffix" {
@@ -52,9 +53,12 @@ resource "google_sql_database_instance" "default" {
 
   settings {
     tier                        = var.tier
+    edition                     = var.edition
     activation_policy           = var.activation_policy
     availability_type           = var.availability_type
     deletion_protection_enabled = var.deletion_protection_enabled
+    connector_enforcement       = local.connector_enforcement
+
     dynamic "backup_configuration" {
       for_each = var.backup_configuration.enabled ? [var.backup_configuration] : []
       content {
@@ -180,8 +184,14 @@ resource "google_sql_database" "additional_databases" {
 }
 
 resource "random_password" "user-password" {
-  length     = 8
-  special    = true
+  length  = 8
+  special = true
+
+  lifecycle {
+    ignore_changes = [
+      special, length
+    ]
+  }
   depends_on = [null_resource.module_depends_on, google_sql_database_instance.default]
 }
 
@@ -190,8 +200,14 @@ resource "random_password" "additional_passwords" {
   keepers = {
     name = google_sql_database_instance.default.name
   }
-  length     = 32
-  special    = true
+  length  = 32
+  special = true
+
+  lifecycle {
+    ignore_changes = [
+      special, length
+    ]
+  }
   depends_on = [null_resource.module_depends_on, google_sql_database_instance.default]
 }