+++ modules/postgresql/read_replica.tf
@@ -37,6 +37,7 @@ resource "google_sql_database_instance" "replicas" {
 
   settings {
     tier                        = lookup(each.value, "tier", var.tier)
+    edition                     = lookup(each.value, "edition", var.edition)
     activation_policy           = "ALWAYS"
     availability_type           = lookup(each.value, "availability_type", var.availability_type)
     deletion_protection_enabled = var.read_replica_deletion_protection_enabled
@@ -65,6 +66,7 @@ resource "google_sql_database_instance" "replicas" {
 
       content {
         query_insights_enabled  = true
+        query_plans_per_minute  = lookup(insights_config.value, "query_plans_per_minute", 5)
         query_string_length     = lookup(insights_config.value, "query_string_length", 1024)
         record_application_tags = lookup(insights_config.value, "record_application_tags", false)
         record_client_address   = lookup(insights_config.value, "record_client_address", false)
@@ -107,4 +109,3 @@ resource "google_sql_database_instance" "replicas" {
     delete = var.delete_timeout
   }
 }
-