+++ modules/postgresql/variables.tf
@@ -54,6 +54,12 @@ variable "tier" {
   default     = "db-f1-micro"
 }
 
+variable "edition" {
+  description = "The edition of the instance, can be ENTERPRISE or ENTERPRISE_PLUS."
+  type        = string
+  default     = null
+}
+
 variable "zone" {
   type        = string
   description = "The zone for the master instance, it should be something like: `us-central1-a`, `us-east1-c`."
@@ -193,6 +199,7 @@ variable "backup_configuration" {
 variable "insights_config" {
   description = "The insights_config settings for the database."
   type = object({
+    query_plans_per_minute  = number
     query_string_length     = number
     record_application_tags = bool
     record_client_address   = bool
@@ -400,6 +407,12 @@ variable "user_deletion_policy" {
 
 variable "enable_random_password_special" {
   description = "Enable special characters in generated random passwords."
+  type        = bool
+  default     = false
+}
+
+variable "connector_enforcement" {
+  description = "Enforce that clients use the connector library"
   type        = bool
   default     = false
 }