+++ modules/safer_mysql/versions.tf
@@ -20,12 +20,12 @@ terraform {
 
     google = {
       source  = "hashicorp/google"
-      version = ">= 4.48.0, < 5.0"
+      version = ">= 4.74.0, < 5.0"
     }
   }
 
   provider_meta "google" {
-    module_name = "blueprints/terraform/terraform-google-sql-db:safer_mysql/v15.2.0"
+    module_name = "blueprints/terraform/terraform-google-sql-db:safer_mysql/v16.1.0"
   }
 
 }